+++ sdk/needle/CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial internal package: re-export get_it and expose `sl`.